Split the fdb-tuple module into its own crate
