+++ Cargo.toml
@@ -2,6 +2,7 @@
 resolver = "2"
 members = [
     "foundationdb",
+    "foundationdb-tuple",
     "foundationdb-sys",
     "foundationdb-gen",
     "foundationdb-bench",
