+++ foundationdb-tuple/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "foundationdb"
+name = "foundationdb-tuple"
 version = "0.9.1"
 authors = [
     "Benjamin Fry <benjaminfry@me.com>",
@@ -10,7 +10,7 @@ edition = "2021"
 rust-version = "1.71"
 
 description = """
-High level client bindings for FoundationDB.
+Fdb tuple impl
 """
 
 documentation = "https://docs.rs/foundationdb"
@@ -22,38 +22,17 @@ keywords = ["foundationdb", "kv"]
 categories = ["database"]
 
 [package.metadata.docs.rs]
-features = ["embedded-fdb-include", "fdb-7_3", "uuid"]
+features = ["uuid"]
 
 [badges]
 codecov = { repository = "foundationdb-rs/foundationdb-rs", branch = "main", service = "github" }
 
 [features]
 default = ["uuid"]
-uuid = ["dep:uuid", "foundationdb-tuple/uuid"]
-num-bigint = ["dep:num-bigint", "foundationdb-tuple/num-bigint"]
-# Use the locally embedded foundationdb fdb_c.h and fdb.options files
-embedded-fdb-include = [
-    "foundationdb-sys/embedded-fdb-include",
-    "foundationdb-gen/embedded-fdb-include",
-]
-fdb-5_1 = ["foundationdb-sys/fdb-5_1", "foundationdb-gen/fdb-5_1"]
-fdb-5_2 = ["foundationdb-sys/fdb-5_2", "foundationdb-gen/fdb-5_2"]
-fdb-6_0 = ["foundationdb-sys/fdb-6_0", "foundationdb-gen/fdb-6_0"]
-fdb-6_1 = ["foundationdb-sys/fdb-6_1", "foundationdb-gen/fdb-6_1"]
-fdb-6_2 = ["foundationdb-sys/fdb-6_2", "foundationdb-gen/fdb-6_2"]
-fdb-6_3 = ["foundationdb-sys/fdb-6_3", "foundationdb-gen/fdb-6_3"]
-fdb-7_0 = ["foundationdb-sys/fdb-7_0", "foundationdb-gen/fdb-7_0"]
-fdb-7_1 = ["foundationdb-sys/fdb-7_1", "foundationdb-gen/fdb-7_1", "serde", "serde_json", "serde_bytes"]
-fdb-7_3 = ["foundationdb-sys/fdb-7_3", "foundationdb-gen/fdb-7_3", "serde", "serde_json", "serde_bytes"]
-tenant-experimental = []
 
-[build-dependencies]
-foundationdb-gen = { version = "0.9.1", path = "../foundationdb-gen", default-features = false }
 
 [dependencies]
-foundationdb-sys = { version = "0.9.1", path = "../foundationdb-sys", default-features = false }
 foundationdb-macros = { version = "0.3.1", path = "../foundationdb-macros" }
-foundationdb-tuple = { version = "0.9.1", path = "../foundationdb-tuple" }
 futures = "0.3.31"
 memchr = "2.7.4"
 rand = { version = "0.8.5", features = ["default", "small_rng"] }
