+++ foundationdb-tuple/src/lib.rs
@@ -3,7 +3,6 @@
 //! The official specification can be found [here](https://github.com/apple/foundationdb/blob/master/design/tuple.md).
 
 mod element;
-pub mod hca;
 mod pack;
 mod subspace;
 mod versionstamp;
