+++ foundationdb/src/lib.rs
@@ -33,7 +33,13 @@ pub mod options;
 ))]
 pub mod tenant;
 mod transaction;
-pub mod tuple;
+mod tuple_ext;
+
+pub mod tuple {
+    pub use foundationdb_tuple::*;
+    pub use crate::tuple_ext::*;
+}
+
 
 #[cfg(any(feature = "fdb-5_1", feature = "fdb-5_2", feature = "fdb-6_0"))]
 pub use crate::cluster::Cluster;
